# project_files/api.yaml
openapi: 3.0.3
info:
  title: Movies API
  version: 1.0.0
  description: |
    API for movies from the last 30 years.
    Provides routes for listing, retrieving, creating, updating, deleting, and analyzing movies.
    Supports filtering, pagination, and statistics grouped by year, genre, or other dimensions.
servers:
  - url: https://dataset-web-api.onrender.com/api
    description: Render (production)
  - url: http://localhost:8000/api
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key required for all operations concerning the movies database
  schemas:
...
    Movie:
      type: object
      required: [id, title]
      properties:
        id:
          type: integer
          example: 101
          description: Unique movie identifier
        title:
          type: string
          example: "Inception"
          description: Movie title
        original_title:
          type: string
          nullable: true
          example: "Inception"
          description: Original title if different
        release_date:
          type: string
          format: date
          example: "2010-07-16"
          description: Release date
        runtime:
          type: integer
          example: 148
          description: Runtime in minutes
        genres:
          type: array
          items:
            type: string
          description: Array of genre names
          example: ["Science Fiction", "Thriller"]
        overview:
          type: string
          nullable: true
          description: Movie plot summary
        budget:
          type: integer
          nullable: true
          example: 160000000
          description: Production budget in USD
        revenue:
          type: integer
          nullable: true
          example: 829895144
          description: Box office revenue in USD
        studios:
          type: array
          items:
            type: string
          description: Array of studio names
          example: ["Warner Bros.", "Legendary Pictures"]
        producers:
          type: array
          items:
            type: string
          description: Array of producer names
          example: ["Emma Thomas", "Christopher Nolan"]
        directors:
          type: array
          items:
            type: string
          description: Array of director names
          example: ["Christopher Nolan"]
        mpa_rating:
          type: string
          nullable: true
          example: "PG-13"
          description: MPA rating
        collection:
          type: string
          nullable: true
          example: "The Dark Knight Collection"
          description: Movie collection/franchise
        poster_url:
          type: string
          format: uri
          nullable: true
          description: Poster image URL
        backdrop_url:
          type: string
          format: uri
          nullable: true
          description: Backdrop image URL
        studio_logos:
          type: array
          items:
            type: string
          description: Array of studio logo URLs
          nullable: true
        studio_countries:
          type: array
          items:
            type: string
          description: Array of studio countries
          nullable: true
        actors:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                nullable: true
              character:
                type: string
                nullable: true
              profile:
                type: string
                format: uri
                nullable: true
          description: Array of up to 10 actors with character and profile info
    PagedMovies:
      type: object
      properties:
        success:
          type: boolean
          example: true
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 25
        data:
          type: array
          items:
            $ref: '#/components/schemas/Movie'
        timestamp:
          type: string
          format: date-time
    StatBucket:
      type: object
      properties:
        key:
          type: string
          example: "2010"
          description: The grouping key (genre, year, etc.)
        count:
          type: integer
          example: 145
          description: Number of movies in this group
    NumericStats:
      type: object
      properties:
        count:
          type: integer
          description: Total count of records
        avg:
          type: number
          format: float
          description: Average value
        min:
          type: number
          description: Minimum value
        max:
          type: number
          description: Maximum value
        sum:
          type: number
          description: Sum of all values
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Movie not found"
        code:
          type: string
          example: "MOVIE_NOT_FOUND"
        timestamp:
          type: string
          format: date-time
security:
  - ApiKeyAuth: []
paths:
  /movies:
    get:
      summary: List movies with basic pagination
      description: |
        Returns a paginated list of movies with basic filtering options.
        This endpoint supports simple pagination without advanced filters.
      tags:
        - Movies
      security: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
          description: Number of movies per page
      responses:
        '200':
          description: Successfully retrieved movies list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedMovies'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /movies/page:
    get:
      summary: Advanced paginated list with comprehensive filters
      description: |
        Returns a paginated list of movies with advanced filtering capabilities.
        Supports filtering by year range, runtime, budget, revenue, genre, rating,
        title search, studios, producers, directors, collections, and actor names.
        All filters can be combined to narrow down search results.
      tags:
        - Movies
      security: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number (starting from 1)
        - in: query
          name: limit
          schema:
            type: integer
            default: 25
            minimum: 1
            maximum: 100
          description: Number of movies per page
        - in: query
          name: yearStart
          schema:
            type: integer
            minimum: 1900
            maximum: 2100
          description: Minimum release year
        - in: query
          name: yearEnd
          schema:
            type: integer
            minimum: 1900
            maximum: 2100
          description: Maximum release year
        - in: query
          name: year
          schema:
            type: integer
            minimum: 1900
            maximum: 2100
          description: Filter by specific release year
        - in: query
          name: runtimeMin
          schema:
            type: integer
            minimum: 0
          description: Minimum runtime in minutes
        - in: query
          name: runtimeMax
          schema:
            type: integer
            minimum: 0
          description: Maximum runtime in minutes
        - in: query
          name: budgetMin
          schema:
            type: integer
            minimum: 0
          description: Minimum budget in USD
        - in: query
          name: budgetMax
          schema:
            type: integer
            minimum: 0
          description: Maximum budget in USD
        - in: query
          name: revenueMin
          schema:
            type: integer
            minimum: 0
          description: Minimum revenue in USD
        - in: query
          name: revenueMax
          schema:
            type: integer
            minimum: 0
          description: Maximum revenue in USD
        - in: query
          name: title
          schema:
            type: string
          description: Filter by partial movie title match (case-insensitive)
        - in: query
          name: genre
          schema:
            type: string
          description: Filter by genre (exact match)
        - in: query
          name: mpaRating
          schema:
            type: string
          description: Filter by MPA rating (e.g., PG-13, R)
        - in: query
          name: studios
          schema:
            type: string
          description: Filter by studio name (case-insensitive)
        - in: query
          name: producers
          schema:
            type: string
          description: Filter by producer name (case-insensitive)
        - in: query
          name: directors
          schema:
            type: string
          description: Filter by director name (case-insensitive)
        - in: query
          name: collection
          schema:
            type: string
          description: Filter by movie collection name
        - in: query
          name: posterUrl
          schema:
            type: string
          description: Filter by poster image URL or partial match
        - in: query
          name: backdropUrl
          schema:
            type: string
          description: Filter by backdrop image URL or partial match
        - in: query
          name: studioLogos
          schema:
            type: string
          description: Filter by studio logo path or reference
        - in: query
          name: studioCountries
          schema:
            type: string
          description: Filter by studio country name
        - in: query
          name: actorNames
          schema:
            type: string
          description: Comma-separated actor names to filter by
      responses:
        '200':
          description: Paginated movie results with applied filters
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 25
                  offset:
                    type: integer
                    example: 0
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Movie'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /movies/random:
    get:
      summary: Get random movies
      description: Returns a random selection of 10 movies from the database.
      tags:
        - Movies
      security: []
      responses:
        '200':
          description: Array of random movies
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Movie'
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /movies/stats:
    get:
      summary: Movie statistics
      description: |
        Returns aggregated movie statistics grouped by the specified dimension.
        Supports both categorical groupings (genre, year, etc.) and numeric
        aggregations (runtime, budget, revenue).
      tags:
        - Movies
      security: []
      parameters:
        - in: query
          name: by
          required: true
          schema:
            type: string
            enum:
              - genre
              - year
              - mpa_rating
              - producers
              - directors
              - studios
              - collection
              - runtime
              - budget
              - revenue
          description: Dimension to group statistics by
      responses:
        '200':
          description: Statistics results
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                      groupedBy:
                        type: string
                        example: "genre"
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/StatBucket'
                      count:
                        type: integer
                        example: 24
                      timestamp:
                        type: string
                        format: date-time
                  - type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                      groupedBy:
                        type: string
                        example: "revenue"
                      data:
                        $ref: '#/components/schemas/NumericStats'
                      timestamp:
                        type: string
                        format: date-time
        '400':
          description: Missing or invalid 'by' parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /movies/getID/{id}:
    get:
      summary: Get movie details by ID
      description: Retrieve detailed information about a specific movie.
      tags:
        - Movies
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Numeric movie ID
      responses:
        '200':
          description: Movie details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Movie'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /movies/post:
    post:
      summary: Create a new movie
      description: |
        Create a new movie record in the database. Requires API key authentication.
        Only the title field is required; all other fields are optional.
      tags:
        - Movies
      operationId: createMovie
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
            example:
              title: "The Quantum Paradox"
              release_date: "2024-06-15"
              runtime: 135
              genres: "Science Fiction; Thriller"
              overview: "A physicist discovers a portal to parallel universes..."
              budget: 75000000
              mpa_rating: "PG-13"
              directors: "Jane Director"
              studios: "Quantum Pictures"
              actor1_name: "John Actor"
              actor1_character: "Dr. Smith"
      responses:
        '201':
          description: Movie created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Movie created successfully"
                  data:
                    $ref: '#/components/schemas/Movie'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /movies/patchID/{id}:
    patch:
      summary: Partially update a movie
      description: |
        Update specific fields of an existing movie. Only provided fields will be updated.
        Requires API key authentication.
      tags:
        - Movies
      operationId: patchMovie
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Movie ID to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
            example:
              revenue: 185000000
              mpa_rating: "R"
              overview: "Updated description with more details..."
      responses:
        '200':
          description: Movie updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Movie partially updated successfully"
                  data:
                    $ref: '#/components/schemas/Movie'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Validation error or no fields provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /movies/deleteID/{id}:
    delete:
      summary: Delete a movie
      description: |
        Remove a movie from the database by ID. Requires API key authentication.
      tags:
        - Movies
      operationId: deleteMovie
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Movie ID to delete
      responses:
        '200':
          description: Movie deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Movie with ID 5001 deleted successfully"
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized - missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'